refactor: decouple demos from sorting_algorithms_n2 directory
