+++ CMakeLists.txt
@@ -39,6 +39,7 @@ set(SRC_DIRS
     src/compression
     demos/data_structures
     demos/trees
+    demos/sorting_algorithms_n2
 )
 
 # Header search paths
